backend/pdf-processing-pipeline/requirements.txt
Pillow==8.4.0
PyPDF2==1.26.0
textwrap3==0.9.2
aiofiles==0.8.0
PyYAML==5.4.1
loguru==0.5.3